# Task5/arkpz-pzpi-22-6-harmaiev-mykhailo-task5/RoadMonitoringSystem/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Road Monitoring System API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Sensors": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Отримує список всіх сенсорів (доступно для User, Operator, Admin).",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Sensor"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sensors"
        ],
        "summary": "Додає новий сенсор (тільки для Admin).",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SensorDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Sensor"
            }
          }
        }
      }
    },
    "/api/Sensors/{id}": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Отримує деталі сенсора за ID (доступно для User, Operator, Admin).",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Sensor"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sensors"
        ],
        "summary": "Оновлює інформацію про сенсор (тільки для Operator, Admin).",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SensorDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Sensors"
        ],
        "summary": "Видаляє сенсор за ID (тільки для Admin).",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Alert": {
      "description": "Представляє сповіщення про критичну ситуацію на дорозі.",
      "required": [
        "alertType",
        "createdDate",
        "isResolved",
        "message",
        "roadSectionID"
      ],
      "type": "object",
      "properties": {
        "alertID": {
          "format": "int32",
          "type": "integer"
        },
        "roadSectionID": {
          "format": "int32",
          "type": "integer"
        },
        "alertType": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "message": {
          "maxLength": 255,
          "minLength": 1,
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "isResolved": {
          "type": "boolean"
        },
        "roadSection": {
          "$ref": "#/definitions/RoadSection"
        }
      },
      "additionalProperties": false
    },
...
    "RoadSection": {
      "description": "Представляє ділянку дороги.",
      "required": [
        "createdDate",
        "name"
      ],
      "type": "object",
      "properties": {
        "roadSectionID": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "location": {
          "maxLength": 255,
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "sensors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Sensor"
          }
        },
        "alerts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Alert"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Sensor": {
      "description": "Представляє сенсор, встановлений на ділянці дороги.",
      "required": [
        "installationDate",
        "roadSectionID",
        "sensorType",
        "status"
      ],
      "type": "object",
      "properties": {
        "sensorID": {
          "format": "int32",
          "type": "integer"
        },
        "roadSectionID": {
          "format": "int32",
          "type": "integer"
        },
        "sensorType": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "installationDate": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "roadSection": {
          "$ref": "#/definitions/RoadSection"
        },
        "sensorData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SensorData"
          }
        }
      },
      "additionalProperties": false
    },
    "SensorData": {
      "description": "Представляє дані, отримані від сенсора.",
      "required": [
        "dataValue",
        "parameter",
        "sensorID",
        "timestamp"
      ],
      "type": "object",
      "properties": {
        "sensorDataID": {
          "format": "int32",
          "type": "integer"
        },
        "sensorID": {
          "format": "int32",
          "type": "integer"
        },
        "parameter": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "dataValue": {
          "format": "double",
          "type": "number"
        },
        "timestamp": {
          "format": "date-time",
          "type": "string"
        },
        "sensor": {
          "$ref": "#/definitions/Sensor"
        }
      },
      "additionalProperties": false
    },
...
    "SensorDto": {
      "type": "object",
      "properties": {
        "roadSectionID": {
          "format": "int32",
          "type": "integer"
        },
        "sensorType": {
          "type": "string"
        },
        "installationDate": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}